PR#7506: Range Locking: Implementation: fix issues detected by CI
Followup to part #2: Make DBT::size and ulen be of type size_t.

+++ utilities/transactions/lock/range/range_tree/util/dbt.h
@@ -54,7 +54,7 @@ DBT *toku_init_dbt_flags(DBT *, uint32_t flags);
 
 void toku_destroy_dbt(DBT *);
 
-DBT *toku_fill_dbt(DBT *dbt, const void *k, uint32_t len);
+DBT *toku_fill_dbt(DBT *dbt, const void *k, size_t len);
 
 DBT *toku_memdup_dbt(DBT *dbt, const void *k, size_t len);
